Add all pages as separate files
